Defer test key generation in anonymous_tokens tests (#44)
* Defer test key generation in anonymous_tokens tests

This change modifies parameterized tests in anonymous_tokens to avoid
calling key generation functions like `GetStrongRsaKeys2048()` directly
in `INSTANTIATE_TEST_SUITE_P`.

Currently, these functions are invoked every time
`testing::InitGoogleTest()` is called. This results in unnecessary
overhead, as the test data is loaded even if the tests are filtered out
by `--gtest_filter`.

More critically, this causes a crash in Android multi-process tests.
These key generation functions depend on
`PathService::Get(DIR_SRC_TEST_DATA_ROOT)` to load test data. When a new
test process is launched via `SpawnMultiProcessTestChild`,
`testing::InitGoogleTest()` is called before `InitAndroidTestPaths()`,
which sets up the required path providers. This leads to a
`NOTREACHED()` failure in `base/base_paths.cc`'s `PathProvider()`.

To fix this, this CL passes function pointers to the key generation
methods as the test parameters and defers the actual invocation of
these functions until the `SetUp()` method of the test fixture is
called. This ensures that the test data is loaded only when the
test is actually run, at which point the PathService is correctly
initialized.

* Remove duplicated